refactor(api): use explicit leaderboard modules
