--- apps/api/src/services/leaderboard.ts
@@ -1,2 +1,0 @@
-export * from './leaderboard-calculation'
-export * from './leaderboard-queries'
